Change read/write commands to operate on arrays of larpix_data
This is a bigger change to larpix_write_data: instead of a single data
object, you pass an array of data objects and each one is written out in
succession. Conveniently, to send just one data object you don't need to
make an array: just pass &data and num_reads=1.