Refactor Controller and SerialPort so Controller never deals with bytes
SerialPort has ``send`` command which accepts a list of packets to send.
The SerialPort converts those packets to bytes and sends them off.

For reading packets, there are commands ``start_listening`` and
``stop_listening`` which open and close the serial port, and ``read``
which reads whatever's in the port's input buffer and returns a list of
packets to the Controller.