Update unit tests to new controller/io scheme
This includes a new FakeIO class that prints "sent" packets and reads
"received" packets from a predefined queue.

It also includes skipping a lot of tests specific to the serial IO
interface and format, which will have to be repurposed as some sort of
test suite for just the SerialPort class.

It also also includes skipping the entire test_simulation.py suite since
the old larpix/simulation.py module was never used for anything and
should be deleted.