Add Logger.READ and Logger.WRITE for direction instead of 0 and 1
Includes the distinction between the hdf5format specification and the
Logger class, i.e. in hdf5format.to_file, an explicit conversion is done
between Logger.READ/WRITE and the HDF5 values 1 and 0. This gives
freedom to both Logger and the hdf5format spec to change the meaning of
READ and WRITE.